+++ src/IotHub/IotHub.Test/ScenarioTests/IotHubDPDeviceTests.ps1
@@ -65,6 +65,11 @@ function Test-AzureRmIotHubDeviceLifecycle
 	$devices = Get-AzIotHubDevice -ResourceGroupName $ResourceGroupName -IotHubName $IotHubName
 	Assert-True { $devices.Count -eq 3}
 
+	# Get device connection string
+	$deviceCS = Get-AzIotHubDCS -ResourceGroupName $ResourceGroupName -IotHubName $IotHubName -DeviceId $device3
+	Assert-True { $deviceCS.DeviceId -eq $device3 }
+	Assert-True { $deviceCS.ConnectionString -eq "HostName=$IotHubName.azure-devices.net;DeviceId=$device3;x509=true" }
+
 	# Update Device
 	$updatedDevice1 = Set-AzIoTHubDevice -ResourceGroupName $ResourceGroupName -IotHubName $IotHubName -DeviceId $device1 -Status 'Disabled' -StatusReason 'Reason1'
 	Assert-True { $updatedDevice1.Id -eq $device1 }
@@ -72,8 +77,20 @@ function Test-AzureRmIotHubDeviceLifecycle
 	Assert-True { $updatedDevice1.StatusReason -eq 'Reason1' }
 
 	# Update iot device to edge device
-	$updatedDevice2 = Set-AzIoTHubDevice -ResourceGroupName $ResourceGroupName -IotHubName $IotHubName -DeviceId $device3 -EdgeEnabled $true
-	Assert-True { $updatedDevice2.Capabilities.IotEdge }
+	$updatedDevice3 = Set-AzIoTHubDevice -ResourceGroupName $ResourceGroupName -IotHubName $IotHubName -DeviceId $device3 -EdgeEnabled $true
+	Assert-True { $updatedDevice3.Capabilities.IotEdge }
+
+	# Set parent device Id
+	$updatedChildDevice = Set-AzIotHubDeviceParent -ResourceGroupName $ResourceGroupName -IotHubName $IotHubName -DeviceId $device1 -ParentDeviceId $device3
+	Assert-False { $updatedChildDevice.Capabilities.IotEdge }
+	Assert-True { $updatedChildDevice.Id -eq $device1 }
+	Assert-True { $updatedChildDevice.Scope -eq $updatedDevice3.Scope }
+
+	# Get parent device Id
+	$parentDevice = Get-AzIotHubDeviceParent -ResourceGroupName $ResourceGroupName -IotHubName $IotHubName -DeviceId $device1
+	Assert-True { $parentDevice.Capabilities.IotEdge }
+	Assert-True { $parentDevice.Id -eq $device3 }
+	Assert-True { $updatedChildDevice.Scope -eq $updatedDevice3.Scope }
 
 	# Get device detail
 	$iotDevice = Get-AzIotHubDevice -ResourceGroupName $ResourceGroupName -IotHubName $IotHubName -DeviceId $device1